refactor data utils
